fixed tests

+++ tests/solidity/values/smallValues.sol
@@ -1,8 +1,8 @@
 contract smallValues {
-  uint256 n = 0;
+  int256 n = 0;
   
   function f (uint256 newn) public {
-    n = newn;
+    n = int256(newn);
   }
   
   function echidna_findSmall() public returns (bool) {
